Rename RestartingState to ConnectingState
