+++ studio/components/layouts/ProjectLayout/ConnectingState.tsx
